test(libs): :white_check_mark: move search engine tests to libs folder
